+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.0.13
-- WIN: Changed to use CefSharp.WinForms. Experimental!
+- WIN: Changed to use CefSharp.WinForms. Improved performance Issue #17
+- WIN: Added support for touch devices. Issue #19
+- WIN/OSX: Added ON_ESC_KEY event. Issue #20
 
 ### 0.0.12
 - WIN: Upgraded CEF to 55.0.0 Issue #16